Fix "Range iterator outside integer range" error
In 32-bit platforms, test #142 fails with a "Range iterator outside
integer range" error. This is probably due to the fact that the
get_start and get_stop methods return an unsigned integer that is
outside the normal range of a signed 32 bit integer. In 64-bits,
this error does not exhibit itself, presumably because the integers
are 64-bit. To fix, instead of using the range operator .. to do the
iteration, we are using an explicit for loop. This ought to make the
error go away.